requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.19
alembic==1.11.1
pydantic==2.1.1
redis==4.6.0
boto3==1.28.17
jwt==1.3.1
pyjwt==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
faker==19.2.0

